+++ src/Network/MPD/Applicative/Database.hs
@@ -53,18 +53,14 @@ find q = Command p ["find" <@> q]
 findAdd :: Query -> Command ()
 findAdd q = Command emptyResponse ["findadd" <@> q]
 
--- | Lists all tags of the specified type.
+-- | List all tags of the specified type of songs that that satisfy the query.
 --
--- Note that the optional artist value is only ever used if the
--- metadata type is 'Album', and is then taken to mean that the albums
--- by that artist be listed.
-list :: Metadata -> Maybe Artist -> Command [Value]
+-- @since 0.10.0.0
+list :: Metadata -> Query -> Command [Value]
 list m q = Command p c
     where
         p = map Value . takeValues <$> getResponse
-        c = case m of
-                Album -> ["list Album" <@> q]
-                _     -> ["list" <@> m]
+        c = ["list" <@> m <++> q]
 
 -- | List all songs and directories in a database path.
 listAll :: Path -> Command [Path]