+++ src/Network/MPD/Commands/Database.hs
@@ -46,10 +46,12 @@ find = A.runCommand . A.find
 findAdd :: MonadMPD m => Query -> m ()
 findAdd = A.runCommand . A.findAdd
 
--- | List all tags of the specified type.
+-- | List all tags of the specified type of songs that that satisfy the query.
+--
+-- @since 0.10.0.0
 list :: MonadMPD m
      => Metadata -- ^ Metadata to list
-     -> Maybe Artist -> m [Value]
+     -> Query -> m [Value]
 list m = A.runCommand . A.list m
 
 -- | List the songs (without metadata) in a database directory recursively.